requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.1.20 
pinecone-client[grpc]==3.2.2
novita-sdk==0.1.2 
pandas==2.2.2
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9 
streamlit==1.35.0
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.32.3
streamlit-agraph==0.0.61 